Speed up saving large projects
Large projects can be very slow to save. I have a 10 MB project file that took about 20 minutes to save. Upon profiling the save operation with `cProfile` it became clear it was spending most of its time in `PBXObjects.__getitem__` and `PBXBuildFile._get_section`. These both perform lookups over that are linear in the size of the project for each call, leading to (at least) an n^2 algorithm. While most of the time we *need* to perform the lookups like this (since people could have modified the project data structures in arbitrary ways) if we make the simplifying assumption that the project won't change during saves, we can greatly speed these operations up by pre-computing a mapping from the ID we care about to the object we're looking for and make each of these calls constant time.

This optimization brings the save time for the same project down to just 2.3 seconds on my machine (4 orders of magnitude faster). This would likely also help in the case of https://github.com/kronenthaler/mod-pbxproj/issues/247.